db-console: move "hot ranges" page to "top ranges"
Previously, the UI referred to high-activity ranges as "hot ranges",
which implied they were necessarily experiencing high load or activity.
This was inadequate because ranges shown may simply be the highest by
some measure (QPS, CPU usage, etc.) without being truly "hot".

To address this, this patch renames the "hot ranges" page and related
UI components to "top ranges" to more accurately reflect that these
are the top-ranked ranges by various metrics rather than necessarily
experiencing high activity.

Release note (ui change): The "Hot Ranges" page in the DB Console has
been renamed to "Top Ranges" to better reflect that it shows the
highest-ranked ranges by various metrics, not necessarily ranges
experiencing high activity.

Fixes #147330
Epic: CRDB-43150
