+++ pkg/ui/workspaces/db-console/src/app.tsx
@@ -58,7 +58,6 @@ import { IndexDetailsPage } from "src/views/databases/indexDetailsPage";
 import Raft from "src/views/devtools/containers/raft";
 import RaftMessages from "src/views/devtools/containers/raftMessages";
 import RaftRanges from "src/views/devtools/containers/raftRanges";
-import HotRangesPage from "src/views/hotRanges/index";
 import JobDetails from "src/views/jobs/jobDetails";
 import JobsPage from "src/views/jobs/jobsPage";
 import KeyVisualizerPage from "src/views/keyVisualizer";
@@ -82,6 +81,7 @@ import SchedulesPage from "src/views/schedules/schedulesPage";
 import SessionDetails from "src/views/sessions/sessionDetails";
 import SQLActivityPage from "src/views/sqlActivity/sqlActivityPage";
 import StatementDetails from "src/views/statements/statementDetails";
+import HotRangesPage from "src/views/topRanges/index";
 import { SnapshotRouter } from "src/views/tracez_v2/snapshotRoutes";
 import TransactionDetails from "src/views/transactions/transactionDetails";
 import "styl/app.styl";
