+++ pkg/ui/workspaces/db-console/src/views/topRanges/hotRanges.module.styl
