+++ pkg/ui/workspaces/db-console/src/views/topRanges/hotRangesFilter.tsx
