+++ pkg/ui/workspaces/db-console/src/views/topRanges/index.tsx
@@ -26,7 +26,7 @@ import {
 } from "src/redux/hotRanges";
 import { selectNodeLocalities } from "src/redux/localities";
 import { performanceBestPracticesHotSpots } from "src/util/docs";
-import { HotRangesFilter } from "src/views/hotRanges/hotRangesFilter";
+import { HotRangesFilter } from "src/views/topRanges/hotRangesFilter";
 
 import ErrorBoundary from "../app/components/errorMessage/errorBoundary";
 
@@ -60,10 +60,10 @@ const HotRangesPage = () => {
 
   return (
     <React.Fragment>
-      <Helmet title="Hot Ranges" />
-      <h3 className="base-heading">Hot Ranges</h3>
+      <Helmet title="Top Ranges" />
+      <h3 className="base-heading">Top Ranges</h3>
       <Text className={cx("hotranges-description")}>
-        The Hot Ranges table shows ranges receiving a high number of reads or
+        The Top Ranges table shows ranges receiving a high number of reads or
         writes. By default, the table is sorted by ranges with the highest QPS
         (queries per second). <br />
         Use this information to
